ses-elo-alt-main/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
haversine==2.8.1
httplib2==0.20.4
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
pyOpenSSL==23.2.0
requests==2.31.0
ssh-import-id==5.11
#ufw==0.36.21
urllib3==2.0.7
Werkzeug==3.0.6
python-dotenv